requirements.txt
requests>=2.31.0
python-amazon-paapi>=1.1.0
